--- src/hypervibes/index.ts
@@ -1,3 +1,0 @@
-export * from './constants';
-export * from './subgraph';
-export * from './contract';
